+++ model_fetch/lib/model_fetch.dart
@@ -1,5 +1,6 @@
 export 'src/abstract_client_filter.dart';
 export 'src/abstract_filter.dart';
-export 'src/lazy_load_bloc.dart';
+export 'src/collection/collection_bloc.dart';
+export 'src/collection/lazy_load_bloc.dart';
 export 'src/load_status.dart';
-export 'src/model_by_id_bloc.dart';
+export 'src/model_by_id/model_by_id_bloc.dart';