+++ model_fetch/lib/src/collection/lazy_load_bloc.dart
@@ -1,0 +1,10 @@
+import 'collection_bloc.dart';
+
+abstract class LazyLoadBloc<T> extends CollectionBloc<T> {
+  LazyLoadBloc({
+    super.totalLimit,
+    super.clientFilters = const [],
+  });
+
+  Future<void> loadMoreIfCan();
+}