+++ model_fetch/lib/src/model_by_id/model_by_id_bloc.dart
@@ -2,7 +2,7 @@ import 'package:meta/meta.dart';
 import 'package:model_interfaces/model_interfaces.dart';
 import 'package:rxdart/rxdart.dart';
 
-import 'load_status.dart';
+import '../load_status.dart';
 
 abstract class ModelByIdBloc<I, T extends WithId<I>> {
   final I id;