+++ model_fetch_firestore/lib/model_fetch_firestore.dart
@@ -1,4 +1,5 @@
-export 'src/lazy_load/firestore_lazy_load_bloc.dart';
-export 'src/lazy_load/firestore_frozen_lazy_load_bloc.dart';
+export 'src/collection/firestore_live_single_document_collection_bloc.dart';
+export 'src/collection/firestore_lazy_load_bloc.dart';
+export 'src/collection/firestore_frozen_lazy_load_bloc.dart';
 export 'src/model_by_id/firestore_live_by_id_bloc.dart';
 export 'src/query_builder.dart';