+++ model_fetch_flutter/lib/model_fetch_flutter.dart
@@ -1,2 +1,2 @@
-export 'src/lazy_load_builder.dart';
+export 'src/collection_bloc_builder.dart';
 export 'src/load_more.dart';