+++ model_fetch/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.3.0
+
+* **BREAKING**: Renamed `LazyLoadState` to `CollectionState`.
+* Added `CollectionBloc` as a superclass of `LazyLoadBloc`.
+
 ## 0.2.0
 
 * **BREAKING**: Renamed `LazyLoadStatus` to `LoadStatus`.