+++ model_fetch_firestore/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.3.0
+
+* Added `FirestoreLiveSingleDocumentCollectionBloc`.
+
 ## 0.2.0
 
 * Added `FirestoreLiveByIdBloc`.