+++ model_fetch_flutter/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.3.0
+
+* **BREAKING**: Renamed `LazyLoadBuilder` to `CollectionBlocBuilder`.
+
 ## 0.2.1
 
 * Support and require Flutter 3.0.